Changes to my branch that are important please believe me.
In all seriousness this is to move the index back into public so that it can deploy locally for debugging and building purpose, I spent some time going through the wildcard manager and commenting out some stuff, I want to try to make it so the new scene can load when you exit out of the menu. But I was having some issues with that.